DevTools - Extracted highlight logic to custom hook. Added highlight support for rendered by elements.
